Move interactive message logic and block formatting to service module
